Debug

+++ models/trainers.py
@@ -262,7 +262,7 @@ class Node_Classif_Exp(pl.LightningModule):
 
     def training_step(self, batch, batch_idx):
         logp = self(batch).permute(0,2,1)
-        loss = self.loss(logp, batch['target'])
+        loss = self.loss(logp, batch['target'].long())
         self.log('train_loss', loss)
         acc = self.accuracy(logp.tensor.rename(None), batch['target'])
         self.log("train_acc", acc)
